tools(vfio-share-gpu): give the delegated-MMIO probe a HIP flavour
The privilege boundary the probe found was measured with CUDA only, while the
design it informs speaks about GPU submission without naming a vendor. HIP has
hipHostRegisterIoMemory, so the same question can be put to ROCm.

The device-side read goes through hipMemcpyDtoH rather than an SM, since this
project has no ROCm kernel-compilation path. That is weaker than the CUDA
flavour and it is enough for the question being asked, which is whether the
registration is accepted at all.
